requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
anthropic>=0.40.0
google-generativeai>=0.8.0
python-dotenv==1.0.1
python-multipart==0.0.20
Pillow>=10.0.0
PyMuPDF>=1.24.0
pydantic>=2.0.0
openpyxl>=3.1.0
fastmcp
